Schedule seed

+++ Personalsystem/Controllers/SchedulesController.cs
@@ -133,9 +133,30 @@ namespace Personalsystem.Controllers
             {
                 return HttpNotFound();
             }
-            ViewBag.DepartmentId = new SelectList(db.Departments, "Id", "Name", schedule.DepartmentId);
-            ViewBag.GroupId = new SelectList(db.DepartmentGroups, "Id", "Name", schedule.GroupId);
-            return View(schedule);
+
+            if (User.Identity.IsAuthenticated)
+            {
+                var userManager = new UserManager<ApplicationUser>(new UserStore<ApplicationUser>(db));
+                var currentUser = userManager.FindById(User.Identity.GetUserId());
+                if (schedule.Department.Bosses.Contains(currentUser) || (schedule.Group != null && schedule.Group.Department.Bosses.Contains(currentUser)))
+                {
+                    ViewBag.DepartmentId = new SelectList(db.Departments.ToList().Where(q => q.Bosses.Contains(currentUser)).ToList(), "Id", "Name", schedule.DepartmentId);
+                    var groups = new SelectList(db.DepartmentGroups.ToList().Where(q => q.Department.Bosses.Contains(currentUser)).ToList(), "Id", "Name", schedule.GroupId).ToList();
+                    groups.Insert(0, new SelectListItem() { Value = "", Text = "All" });
+                    ViewBag.GroupId = groups;
+
+                    return View(schedule);
+                }
+                else
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
+                }
+
+            }
+            else
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
+            }
         }
 
         // POST: Schedules/Edit/5
@@ -145,14 +166,39 @@ namespace Personalsystem.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Id,StartTime,EndTime,DepartmentId,GroupId")] Schedule schedule)
         {
-            if (ModelState.IsValid)
+            var userManager = new UserManager<ApplicationUser>(new UserStore<ApplicationUser>(db));
+            var currentUser = userManager.FindById(User.Identity.GetUserId());
+            if (ModelState.IsValid && User.Identity.IsAuthenticated)
             {
-                db.Entry(schedule).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (schedule.Id == 0)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Schedule dbSchedule = db.Schedules.Find(schedule.Id);
+                if (dbSchedule == null)
+                {
+                    return HttpNotFound();
+                }
+                if (dbSchedule.Department.Bosses.Contains(currentUser) || (dbSchedule.Group != null && dbSchedule.Group.Department.Bosses.Contains(currentUser)))
+                {
+                    db.Entry(dbSchedule).State = EntityState.Detached;
+
+                    db.Entry(schedule).State = EntityState.Modified;
+                    db.SaveChanges();
+
+                    return RedirectToAction("Index", "Schedules", new { id = schedule.DepartmentId });
+                }
+                else
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
+                }
             }
-            ViewBag.DepartmentId = new SelectList(db.Departments, "Id", "Name", schedule.DepartmentId);
-            ViewBag.GroupId = new SelectList(db.DepartmentGroups, "Id", "Name", schedule.GroupId);
+
+            ViewBag.DepartmentId = new SelectList(db.Departments.ToList().Where(q => q.Bosses.Contains(currentUser)).ToList(), "Id", "Name", schedule.DepartmentId);
+            var groups = new SelectList(db.DepartmentGroups.ToList().Where(q => q.Department.Bosses.Contains(currentUser)).ToList(), "Id", "Name", schedule.GroupId).ToList();
+            groups.Insert(0, new SelectListItem() { Value = "", Text = "All" });
+            ViewBag.GroupId = groups;
+
             return View(schedule);
         }
 
@@ -168,7 +214,21 @@ namespace Personalsystem.Controllers
             {
                 return HttpNotFound();
             }
-            return View(schedule);
+            if (User.Identity.IsAuthenticated)
+            {
+                var userManager = new UserManager<ApplicationUser>(new UserStore<ApplicationUser>(db));
+                var currentUser = userManager.FindById(User.Identity.GetUserId());
+                if (schedule.Department.Bosses.Contains(currentUser) || (schedule.Group != null && schedule.Group.Department.Bosses.Contains(currentUser)))
+                {
+                    return View(schedule);
+                }
+                else
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
+                }
+
+            }
+            return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
         }
 
         // POST: Schedules/Delete/5
@@ -176,10 +236,32 @@ namespace Personalsystem.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (id == 0)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             Schedule schedule = db.Schedules.Find(id);
-            db.Schedules.Remove(schedule);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (schedule == null)
+            {
+                return HttpNotFound();
+            }
+            if (User.Identity.IsAuthenticated)
+            {
+                var userManager = new UserManager<ApplicationUser>(new UserStore<ApplicationUser>(db));
+                var currentUser = userManager.FindById(User.Identity.GetUserId());
+                if (schedule.Department.Bosses.Contains(currentUser) || (schedule.Group != null && schedule.Group.Department.Bosses.Contains(currentUser)))
+                {
+                    var departmentId = schedule.DepartmentId;
+                    db.Schedules.Remove(schedule);
+                    db.SaveChanges();
+                    return RedirectToAction("Index", "Schedules", new { id = departmentId });
+                }
+                else
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
+                }
+            }
+            return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
         }
 
         protected override void Dispose(bool disposing)
